renamed certain methods/modules for url sanitization
