+++ code/steps/sanitize_data/tests/test_url_sanitization.py
@@ -4,10 +4,10 @@ import logging
 import sys
 
 from sanitization_code.url_regex import url_regex
-from sanitization_code.url_sanitization.sanitize_urls import *
+from sanitization_code.url_sanitization.sanitize_url import *
 from sanitization_code.url_sanitization.url_sanitization_params import num_threads_default, requests_timeout_default, retry_after_default, max_attempts_default
 from sanitization_code.url_sanitization.parallelize_url_sanitization import *
-from sanitization_code.url_sanitization.sanitize_url_data import *
+from sanitization_code.url_sanitization.get_sanitized_urls_for_update import *
 
 logging.basicConfig(
                     stream = sys.stdout, 
@@ -66,26 +66,26 @@ def test_assign_string_condition():
     assert assign_string_condition(test_string_3, test_string_3_matches) == 'String contains no URLs'
     assert assign_string_condition(test_string_4, test_string_4_matches) == 'String contains multiple URLs'
 
-def test_sanitize_urls_with_no_urls():
+def test_sanitize_url_with_no_urls():
     test_string = ''
 
-    assert sanitize_urls(test_string, requests_timeout_default, retry_after_default, max_attempts_default) == {'raw_string': '', 'condition': 'String contains no URLs', 'URLs': []}
+    assert sanitize_url(test_string, requests_timeout_default, retry_after_default, max_attempts_default) == {'raw_string': '', 'condition': 'String contains no URLs', 'URLs': []}
 
-def test_sanitize_urls_with_only_one_url():
+def test_sanitize_url_with_only_one_url():
     test_string_1 = 'https://www.w3.org/Addressing/URL/url-spec.txt'
     test_string_2 = 'https://www.w3.org/Addressing/URL/url-spec.txtasdf' # additional characters that invalidate full URL match
     test_string_3 = 'URL Specification https://www.w3.org/Addressing/URL/url-spec.txt' # extra characters outside of URL
     test_string_4 = 'URL Specification https://www.w3.org/Addressing/URL/url-spec.txtasdf'
 
-    assert sanitize_urls(test_string_1, requests_timeout_default, retry_after_default, max_attempts_default) == {'raw_string': 'https://www.w3.org/Addressing/URL/url-spec.txt', 'condition': 'String is URL', 'URLs': [{'URL': 'https://www.w3.org/Addressing/URL/url-spec.txt', 'root_URL': 'https://www.w3.org', 'URL_status': 200, 'root_URL_status': 200}]}
-    assert sanitize_urls(test_string_2, requests_timeout_default, retry_after_default, max_attempts_default) == {'raw_string': 'https://www.w3.org/Addressing/URL/url-spec.txtasdf', 'condition': 'String is URL', 'URLs': [{'URL': 'https://www.w3.org/Addressing/URL/url-spec.txtasdf', 'root_URL': 'https://www.w3.org', 'URL_status': 300, 'root_URL_status': 200}]}
-    assert sanitize_urls(test_string_3, requests_timeout_default, retry_after_default, max_attempts_default) == {'raw_string': 'URL Specification https://www.w3.org/Addressing/URL/url-spec.txt', 'condition': 'String is not URL but contains one', 'URLs': [{'URL': 'https://www.w3.org/Addressing/URL/url-spec.txt', 'root_URL': 'https://www.w3.org', 'URL_status': 200, 'root_URL_status': 200}]}
-    assert sanitize_urls(test_string_4, requests_timeout_default, retry_after_default, max_attempts_default) == {'raw_string': 'URL Specification https://www.w3.org/Addressing/URL/url-spec.txtasdf', 'condition': 'String is not URL but contains one', 'URLs': [{'URL': 'https://www.w3.org/Addressing/URL/url-spec.txtasdf', 'root_URL': 'https://www.w3.org', 'URL_status': 300, 'root_URL_status': 200}]}
+    assert sanitize_url(test_string_1, requests_timeout_default, retry_after_default, max_attempts_default) == {'raw_string': 'https://www.w3.org/Addressing/URL/url-spec.txt', 'condition': 'String is URL', 'URLs': [{'URL': 'https://www.w3.org/Addressing/URL/url-spec.txt', 'root_URL': 'https://www.w3.org', 'URL_status': 200, 'root_URL_status': 200}]}
+    assert sanitize_url(test_string_2, requests_timeout_default, retry_after_default, max_attempts_default) == {'raw_string': 'https://www.w3.org/Addressing/URL/url-spec.txtasdf', 'condition': 'String is URL', 'URLs': [{'URL': 'https://www.w3.org/Addressing/URL/url-spec.txtasdf', 'root_URL': 'https://www.w3.org', 'URL_status': 300, 'root_URL_status': 200}]}
+    assert sanitize_url(test_string_3, requests_timeout_default, retry_after_default, max_attempts_default) == {'raw_string': 'URL Specification https://www.w3.org/Addressing/URL/url-spec.txt', 'condition': 'String is not URL but contains one', 'URLs': [{'URL': 'https://www.w3.org/Addressing/URL/url-spec.txt', 'root_URL': 'https://www.w3.org', 'URL_status': 200, 'root_URL_status': 200}]}
+    assert sanitize_url(test_string_4, requests_timeout_default, retry_after_default, max_attempts_default) == {'raw_string': 'URL Specification https://www.w3.org/Addressing/URL/url-spec.txtasdf', 'condition': 'String is not URL but contains one', 'URLs': [{'URL': 'https://www.w3.org/Addressing/URL/url-spec.txtasdf', 'root_URL': 'https://www.w3.org', 'URL_status': 300, 'root_URL_status': 200}]}
 
-def test_sanitize_urls_with_multiple_URLs():
+def test_sanitize_url_with_multiple_URLs():
     test_string = 'https://www.w3.org/Addressing/URL/url-spec.txtasdf https://gist.github.com/dperini/729294 https://miguendes.me/how-to-check-if-a-string-is-a-valid-url-in-python'
 
-    assert sanitize_urls(test_string, requests_timeout_default, retry_after_default, max_attempts_default) == {
+    assert sanitize_url(test_string, requests_timeout_default, retry_after_default, max_attempts_default) == {
         'raw_string': 'https://www.w3.org/Addressing/URL/url-spec.txtasdf https://gist.github.com/dperini/729294 https://miguendes.me/how-to-check-if-a-string-is-a-valid-url-in-python',
         'condition': 'String contains multiple URLs', 
         'URLs': [
@@ -101,7 +101,7 @@ def test_sample_urls_from_within_service():
                 https://www.maxhigbee.org
             '''  
 
-    assert sanitize_urls(urls, requests_timeout_default, retry_after_default, max_attempts_default) == {
+    assert sanitize_url(urls, requests_timeout_default, retry_after_default, max_attempts_default) == {
         'raw_string': urls,
         'condition': 'String contains multiple URLs', 
         'URLs': [
@@ -157,13 +157,12 @@ def test_get_sanitized_urls_for_update():
         key_vals = ['1','2'],
         source_table = 'source_tbl',
         source_column = 'source_col',
-        infokind = 'url',
         logger = logger
     )
 
     # only the unclean url should be sanitized and included in sanitized_urls_json
     assert sanitized_urls_json == [{
         'id': '2',
-        'url': 'https://www.kidsinmotionclinic.org'
+        'URL': 'https://www.kidsinmotionclinic.org'
     }]
